test(e2e): relocate typecheck tests
