+++ AGENTS.md
@@ -136,7 +136,7 @@ Rules that `lint:fix` cannot fix:
 - Strict TypeScript configuration
 - Use `pnpm typecheck` to verify types
 - Configuration files: `tsconfig.base.json`, `tsconfig.build.json`, `tsconfig.check.json`
-- Root `pnpm typecheck` excludes `test/e2e`, `test/browser`, `test/typescript`, `docs`, and `examples` (see `tsconfig.check.json`); type errors there will not surface from the root command
+- Root `pnpm typecheck` excludes `test/e2e`, `test/browser`, `docs`, and `examples` (see `tsconfig.check.json`); type errors there will not surface from the root command
 - Root typecheck does not cover the UI client Vue code; when changing `packages/ui/client`, also run `pnpm -C packages/ui typecheck:client`
 
 ### Code Quality
