+++ test/e2e/package.json
@@ -7,6 +7,8 @@
   },
   "scripts": {
     "test": "vitest",
+    "types": "vitest --root typecheck --typecheck.only --typecheck.allowJs --run",
+    "tsc": "tsc --project typecheck/tsconfig.json --watch --pretty false --noEmit",
     "stacktraces": "vitest --root=./fixtures/stacktraces --watch=false",
     "native": "vitest --root=./fixtures/no-module-runner --watch=false --no-isolate --max-workers=1"
   },
@@ -50,6 +52,7 @@
     "vite": "latest",
     "vitest": "workspace:*",
     "vitest-sonar-reporter": "catalog:",
+    "vue-tsc": "^3.3.11",
     "ws": "catalog:"
   }
 }
