+++ test/e2e/typecheck/failing/js-fail.test-d.js
