+++ test/e2e/typecheck/failing/node-types.test-d.ts
