+++ test/e2e/typecheck/failing/only.test-d.ts
