+++ test/e2e/typecheck/fixtures/dynamic-title/test/dynamic-title.test-d.ts
