+++ test/e2e/typecheck/fixtures/source-error/src/not-ok.ts
