+++ test/e2e/typecheck/fixtures/source-error/test/ok.test-d.ts
