+++ test/e2e/typecheck/fixtures/source-error/tsconfig.json
