+++ test/e2e/typecheck/fixtures/source-error/vite.config.ts
