+++ test/e2e/typecheck/fixtures/typecheck-crash/fake-tsc.mjs
