+++ test/e2e/typecheck/fixtures/typecheck-crash/test/ok.test-d.ts
