+++ test/e2e/typecheck/fixtures/typecheck-crash/tsconfig.json
