+++ test/e2e/typecheck/test-d/expect-extend.test-d.ts
