+++ test/e2e/typecheck/test-d/js.test-d.js
