+++ test/e2e/typecheck/test-d/nested-suite1.test-d.ts
