+++ test/e2e/typecheck/test-d/nested-suite2.test-d.ts
