+++ test/e2e/typecheck/test-d/test.test-d.ts
