+++ test/e2e/typecheck/test/vitest.custom.config.ts
