+++ test/e2e/typecheck/test/vitest.empty.config.ts
