+++ test/e2e/typecheck/tsconfig.custom.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../tsconfig.build.json",
+  "extends": "../../../tsconfig.build.json",
   "include": [
     "**/fail.test-d.ts"
   ],
