+++ test/e2e/typecheck/tsconfig.fails.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../tsconfig.build.json",
+  "extends": "../../../tsconfig.build.json",
   "compilerOptions": {
     "types": []
   },
