+++ test/e2e/typecheck/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../tsconfig.build.json",
+  "extends": "../../../tsconfig.build.json",
   "include": [
     "./**/*.ts",
     "./**/*.js"
