--- test/typescript/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "@vitest/test-typescript",
-  "type": "module",
-  "scripts": {
-    "test": "vitest --test-timeout 60000",
-    "types": "vitest --typecheck.only --typecheck.allowJs --run",
-    "tsc": "tsc --watch --pretty false --noEmit"
-  },
-  "dependencies": {
-    "vitest": "workspace:*"
-  },
-  "devDependencies": {
-    "typescript": "^5.9.3",
-    "vue-tsc": "^3.3.11"
-  }
-}
