+++ tsconfig.check.json
@@ -17,7 +17,6 @@
     "./docs/**",
     "./bench/**",
     "./test/e2e/**",
-    "./test/typescript/**",
     "./test/browser/**",
     "**/coverage/fixtures/**",
     "./test/watch/fixtures/**"
